feat (front): finalizando o modal
