+++ front/src/renderer/src/views/components/DatePicker/index.tsx
@@ -1,36 +1,39 @@
+import { capitalizeFirstLetter } from '@renderer/app/utils/capitalizeFirstLetter';
+import { format } from 'date-fns';
 import { ptBR } from 'date-fns/locale';
-import { DayPicker } from 'react-day-picker';
-import { StyledDayPickerWrapper } from './styles'; // Corrigido o caminho do módulo
+import { StyledDayPicker, StyledSpan } from './styles';
 
 interface DatePickerProps {
   value: Date;
   onChange?(date: Date): void;
 }
 
-export function StyledDatePicker({ value, onChange }: DatePickerProps) {
-
+export default function DatePicker({ value, onChange }: DatePickerProps) {
   return (
-    <StyledDayPickerWrapper>
-      <DayPicker
-        locale={ptBR}
-        selected={value}
-        mode="single"
-        onSelect={(date) => onChange?.(date ?? new Date())}
-        className="rdp-root"
-        classNames={{
-          caption: 'rdp-caption', // Estilização do cabeçalho
-          nav: 'rdp-nav', // Botões de navegação
-          nav_button_previous: 'rdp-nav_button_previous',
-          nav_button_next: 'rdp-nav_button_next',
-          head_cell: 'rdp-head_cell', // Cabeçalho dos dias da semana
-          weekday: 'rdp-weekday', // Estilo dos dias da semana
-          month_grid: 'rdp-month_grid', // Estilo do grid de dias
-          day: 'rdp-day', // Estilo de um dia
-          day_today: 'rdp-day_today', // Estilo do dia atual
-          day_selected: 'rdp-day_selected', // Estilo do dia selecionado
-          day_button: 'rdp-day_button', // Botões dos dias
-        }}
-      />
-    </StyledDayPickerWrapper>
+    <StyledDayPicker
+      locale={ptBR}
+      selected={value}
+      mode="single"
+      onSelect={(date) => onChange?.(date ?? new Date())}
+      classNames={{
+        caption: 'caption',
+        nav: 'nav',
+        nav_button_previous: 'nav_button',
+        nav_button_next: 'nav_button',
+        head_cell: 'head_ceil',
+        button: 'button',
+        day_today: 'day_today',
+        day_selected: 'day_selected',
+      }}
+      formatters={{
+        formatCaption: (date, options) => {
+          return (
+            <StyledSpan>
+              {capitalizeFirstLetter(format(date, 'LLLL yyyy', options))}
+            </StyledSpan>
+          );
+        },
+      }}
+    />
   );
 }
