+++ front/src/renderer/src/views/components/DatePicker/styles.ts
@@ -1,77 +1,67 @@
+import { DayPicker } from 'react-day-picker';
 import styled from 'styled-components';
 
-export const StyledDayPickerWrapper = styled.div`
-  background-color: #fff; /* Cor de fundo clara */
-  padding: 2rem; /* Espaçamento interno */
-  border-radius: 0.5rem; /* Bordas arredondadas */
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /* Sombra suave */
-  display: inline-block; /* Garante que o tamanho seja adequado ao conteúdo */
+import 'react-day-picker/dist/style.css';
 
-  .rdp-caption {
+export const StyledSpan = styled.span`
+  letter-spacing: -0.408px;
+  font-weight: 500;
+  color: #212529;
+`;
+
+export const StyledDayPicker = styled(DayPicker)`
+  .caption {
     display: flex;
+    align-items: center;
     justify-content: space-between;
+  }
+
+  .nav {
+    display: flex;
+    gap: 0.4rem;
+  }
+
+  .nav_button {
+    color: #222 !important;
+    display: flex;
     align-items: center;
-  }
-
-  .rdp-nav {
-    display: flex;
-    gap: 0.25rem;
-  }
-
-  .rdp-nav_button_previous,
-  .rdp-nav_button_next {
-    color: #0d9488;
-    display: flex;
     justify-content: center;
-    align-items: center;
-    background-color: transparent !important;
+    border: none !important ;
+    background: transparent !important;
   }
 
-  .rdp-head_cell {
+  .head_ceil {
     text-transform: uppercase;
-    font-size: 0.75rem;
-    color: #6b7280;
+    font-size: 1.2rem;
+    color: ${({ theme }) => theme.colors.black.main};
     font-weight: 500;
-    padding-top: 0.25rem;
-    padding-bottom: 0.5rem;
+    padding-top: 0.4rem;
+    padding-bottom: 0.8rem;
   }
 
-  .rdp-day_button {
-    background: transparent;
-    margin: 0.5rem;
-    justify-content: center;
-    align-items: center;
-    display: flex;
-    padding: 6px;
-    width: 40px;
-
-    border: var(--rdp-day_button-border);
+  .button {
+    color: ${({ theme }) => theme.colors.gray.main};
+    cursor: pointer;
+    width: 4rem;
+    height: 4rem;
+    border-radius: 50%;
 
     &:hover {
-      background-color: ${({ theme }) =>
-        theme.colors.orange.main}; /* Cor ao passar o mouse */
-      border-radius: 4px;
+      background-color: ${({ theme }) => theme.colors.orange.lighter};
+      color: #fff;
     }
   }
 
-  .rdp-nav {
-    display: flex;
-    align-items: center;
-    margin-left: auto;
-    justify-content: end;
+  .day_today {
+    background: ${({ theme }) => theme.colors.gray.lighter} !important;
+    color: ${({ theme }) => theme.colors.black.light};;
+    font-weight: 700;
   }
 
-  .rdp-button_previous,
-  .rdp-button_next {
-    margin: 0 0.25rem;
-    background: transparent;
-    border: none;
-    cursor: pointer;
-  }
-
-  .rdp-weekday {
-    padding: 0.75rem 1rem; /* Aumenta o espaçamento interno dos dias da semana */
-    text-align: center;
-    font-size: 1.5rem;
+  .day_selected {
+    background: ${({ theme }) => theme.colors.orange.light} !important;
+    color: #fff !important;
+    font-weight: 600;
   }
 `;
+
