+++ front/src/renderer/src/views/components/Popover/index.tsx
@@ -1,22 +1,48 @@
 import * as RdxPopover from '@radix-ui/react-popover';
+import { StyledRdxPopoverContent, StyledRdxPopoverTrigger } from './styles';
 
 function PopoverRoot({ children }: { children: React.ReactNode }) {
   return <RdxPopover.Root>{children}</RdxPopover.Root>;
 }
 
-function PopoverTrigger({ children }: { children: React.ReactNode }) {
-  return <RdxPopover.Trigger asChild>{children}</RdxPopover.Trigger>;
+interface PopoverTriggerProps {
+  children: React.ReactNode;
+  asChild?: boolean;
+}
+
+function PopoverTrigger({ children, asChild }: PopoverTriggerProps) {
+  return (
+    <StyledRdxPopoverTrigger asChild={asChild}>
+      {children}
+    </StyledRdxPopoverTrigger>
+  );
 }
 
 interface PopoverContentProps {
   children: React.ReactNode;
-  className?: string;
+  side?: 'top' | 'right' | 'bottom' | 'left';
+  sideOffset?: number;
+  align?: 'center' | 'start' | 'end';
+  asChild?: boolean;
 }
 
-function PopoverContent({ children, className }: PopoverContentProps) {
+function PopoverContent({
+  children,
+  side,
+  sideOffset,
+  align,
+  asChild,
+}: PopoverContentProps) {
   return (
     <RdxPopover.Portal>
-      <RdxPopover.Content className={className}>{children}</RdxPopover.Content>
+      <StyledRdxPopoverContent
+        side={side}
+        sideOffset={sideOffset}
+        align={align}
+        asChild={asChild}
+      >
+        {children}
+      </StyledRdxPopoverContent>
     </RdxPopover.Portal>
   );
 }
